requirements.txt
annotated-types==0.7.0
anyio==4.8.0
audioop-lts==0.2.1
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
elevenlabs==1.50.7
ffmpeg==1.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
lottie==0.7.1
moviepy==2.1.2
numpy==2.2.2
opencv-python==4.11.0.86
pillow==10.4.0
proglog==0.1.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tinycss2==1.4.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
webencodings==0.5.1
websockets==14.2
natsort==8.4.0
selenium==4.29.0